pyproject.toml
[project]
name = "aws-community-day-demo"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Lingxian Kong", email = "anlin.kong@gmail.com"},
]
dependencies = [
    "python-dotenv>=1.0.1",
    "jupyter>=1.0.0",
    "boto3>=1.34.153",
    "langchain-aws>=0.1.15",
    "langchain>=0.2.12",
    "pyboxen>=1.3.0",
    "langchain-community>=0.2.11",
    "langchain-anthropic>=0.1.22",
    "aws-cdk-lib>=2.151.0",
    "aws-cdk-aws-lambda-python-alpha>=2.151.0a0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
